OBPIH-7864 Fixes after review
